add my orders history

+++ src/components/orderCard/index.jsx
@@ -1,21 +1,32 @@
 import { XMarkIcon } from "@heroicons/react/24/solid";
+import { render } from "react-dom";
 
-const OrderCard = props => {
-  const { id,title,imageUrl, price, handleDelete} = props
+const OrderCard = (props) => {
+  const { id, title, imageUrl, price, handleDelete } = props;
+  let renderXMarkIcon
+
+  if (handleDelete) {
+    renderXMarkIcon = <XMarkIcon 
+    onClick={() => handleDelete(id)}
+    className="h-6 w-6 cursor-pointer">
+    </XMarkIcon>
+  }
+
   return (
     <div className="flex justify-between items-center mb-3">
       <div className="flex items-center gap-2">
-        <figure className='w-20 h-20'>
-          <img className ='w-full h-full rounded-lg object-cover' src={imageUrl} alt={title} />
+        <figure className="w-20 h-20">
+          <img
+            className="w-full h-full rounded-lg object-cover"
+            src={imageUrl}
+            alt={title}
+          />
         </figure>
-        <p className='text-sm font-light '>{title}</p>
+        <p className="text-sm font-light ">{title}</p>
       </div>
       <div className='flex items-center gap-2'>
         <p className='text-sm font-medium' >{price}</p>
-        <XMarkIcon 
-        onClick={() => handleDelete(id)}
-        className="h-6 w-6 cursor-pointer">
-        </XMarkIcon>
+        {renderXMarkIcon}
       </div>
     </div>
   );
